Moved correct_input to helper function.
